rename the list_tags to just list
